authentication, users implemented with playlists features

+++ src/services/postgres/playlists.js
@@ -2,7 +2,7 @@ const {Pool} = require('pg');
 const {nanoid} = require('nanoid');
 const InvariantError = require('../../exceptions/InvariantError');
 const NotFoundError = require('../../exceptions/NotFoundError');
-const AuthenticationError = require('../../exceptions/AuthenticationError');
+const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class PlaylistsService {
     constructor(){
@@ -38,10 +38,10 @@ class PlaylistsService {
         return result.rows;
     }
 
-    async deletePlaylist(id){
+    async deletePlaylist(id, owner){
         const query = {
-            text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
-            values: [id],
+            text: 'DELETE FROM playlists WHERE id = $1 AND owner = $2 RETURNING id',
+            values: [id, owner],
         };
 
         const result = await this._pool.query(query);
@@ -52,6 +52,7 @@ class PlaylistsService {
     }
 
     async addPlaylistSong(playlistId, songId){
+        await this.verifySong(songId);
         const id = `playlist_songs-${nanoid(16)}`;
         const query = {
             text: 'INSERT INTO playlist_songs VALUES($1, $2, $3) RETURNING id',
@@ -68,35 +69,35 @@ class PlaylistsService {
 
     async getPlaylistSong(playlistId){
         const queryPlaylist = {
-            text: `SELECT p.id, p.name, u.username
-            FROM playlists p, users u
-            WHERE p.id = $1 AND p.owner = u.id`,
+            text: `SELECT playlists.id, playlists.name, users.username
+            FROM playlists, users
+            WHERE playlists.id = $1 AND playlists.owner = users.id`,
             values: [playlistId],
         };
         
-        const resultPlaylist = await this._pool.query(queryPlaylist);
+        const resultPlaylist = await this._pool.query(queryPlaylist);        
 
         if(resultPlaylist.rowCount <= 0){
             throw new NotFoundError('Playlist tidak ditemukan.');
         }
 
         const querySong = {
-            text: `SELECT s.id, s.title, s.performer
-            FROM songs s
-            LEFT JOIN playlist_songs p ON p.song_id = s.id
-            WHERE p.playlist_id = $1
-            GROUP BY p.playlist_id`,
+            text: `SELECT songs.id, songs.title, songs.performer
+            FROM songs
+            LEFT JOIN playlist_songs ON playlist_songs.song_id = songs.id
+            WHERE playlist_songs.playlist_id = $1`,
             values: [playlistId],
         };
-
+        
         const resultSong = await this._pool.query(querySong);
-
+       
         resultPlaylist.rows[0].songs = resultSong.rows;
 
         return resultPlaylist.rows[0];
     }
 
     async deletePlaylistSong(playlistId, songId){
+        await this.verifySong(songId);        
         const query = {
             text: 'DELETE FROM playlist_songs WHERE playlist_id = $1 AND song_id = $2 RETURNING id',
             values: [playlistId, songId],
@@ -106,6 +107,36 @@ class PlaylistsService {
 
         if(result.rowCount <= 0){
             return new NotFoundError('Gagal menghapus lagu. Lagu tidak ditemukan');
+        }        
+    }
+
+    async verifyPlaylistOwner(id, owner){
+        const query = {
+            text: 'SELECT * FROM playlists WHERE id = $1',
+            values: [id],
+        };
+
+        const result = await this._pool.query(query);
+        if (!result.rows.length) {
+            throw new NotFoundError('Playlist tidak ditemukan');
+        }
+
+        const playlist = result.rows[0];
+        if (playlist.owner !== owner) {
+            throw new AuthorizationError('Anda tidak berhak mengakses resource ini');
+        }
+    }
+
+    async verifySong(songId){
+        const query = {
+            text: 'SELECT * FROM songs WHERE id = $1',
+            values: [songId],
+        };
+
+        const result = await this._pool.query(query);
+
+        if(result.rowCount <= 0){
+            throw new NotFoundError('Lagu tidak ditemukan');
         }        
     }
 }
